Edits and Adjustments
userDetail: messed with the delete button a little.
userEdit: changed "Back to users index" to be "Back to User Detail"
header: Added a currently non-functional link to signOut
main: Altering '/' route to return the user's role and add it to the cookie.
Altering userDetail/:userID to try to stop non-authenticated users from editing the user, adding a pet, or deleting the user.