#3 implement authentication
/userDetail/:userID page currently checks if token of the currently signed in person is the same as the user being viewed or if the person viewing the information is an admin.

If neither of these things are true, sets pathstatus and buttonStatus to 'disabled'. Later these variables will be passed to the website to disable the viewer's ability to edit the user, add a pet, or delete the user.

If the viewer IS the user, or IS an admin, then those variables are set to enabled and will allow those actions to continue.